Promote cpu reference accuracy to float32

+++ tests/accuracy_utils.py
@@ -76,7 +76,7 @@ def to_reference(inp, upcast=False):
         return None
     ref_inp = inp
     if upcast:
-        ref_inp = ref_inp.to(torch.float64)
+        ref_inp = ref_inp.to(torch.float32)
     if TO_CPU:
         ref_inp = ref_inp.to("cpu")
     return ref_inp
